Distributed: determine bind address from ENV["SSH_CONNECTION"]
If the worker has multiple IP interfaces and/or multiple IP addresses
addresses, it seems preferable (and safer) that it binds by default to
the same IP address that ssh has already successfully used to reach
it. Popular sshd implementations provide that IP address to the worker
in the SSH_CONNECTION environment variable.

This patch also replaces use of the deprecated getipaddr() function
with getipaddrs(), thereby removing IPv4 bias.